+++ sample/resourcetest/attributes/default.rb
@@ -1,0 +1,1 @@
+default['iis']['docroot'] = 'C:/inetpub/wwwroot'